Create FizzBuzz game
To print the numbers between 1 to 10. If a number is divisible by 3, instead of printing that, the program will print "Fizz". If a number is divisible by 5, instead of printing that, the program will print "Buzz". And when a number is divisible by both 3 and 5, the program will print "FizzBuzz".